Rename config files to be ready for the refactoring
* Renamed cfnconfig.py to pcluster_config.py
* Renamed config_sanitiy.py to validators.py
